Remove codes of broadcasting peer ids
The broadcasting mechanism was added to the project because there was
some message loss issues in the `ts-guard` project. As those issues are
fixed now, the broadcasting of peer ids are neither needed nor suggested,
because this broadcasting itself introduces a bunch of new bugs in the
`ts-guard` project, causing the guards to lose their connections with
each other after a period of time.

Also mark `broadcastPeerIds` in `relay` as deprecated.