Increase hop timeout to 10 minutes
Due to the duration of guards messaging scenario, the default hop timeout
of 30 seconds is low and may cause some issues in the messaging between
guards.

Please note that this value works for now, but in the future and with
different number of guards and relays and their different networks (in
terms of locality), this may need notable changes.