Increase circuit relay stream limits of relay
Increase circuit relay stream limits to 1000. The 1000 limit may be an
overkill for the service, but in order to prevent possible stream limit
overflow issues, we use it until we make sure a smaller number would be
acceptable, too.

Please note that the way we increase circuit relay stream limits is not
ideal, but while libp2p apis doesn't support it, we must stick to these
ugly solutions.